Монтана_Montana_蒙大拿/Русский/бот/requirements.txt
# Junona Bot - зависимости
# pip install -r requirements.txt

# Telegram
python-telegram-bot>=21.0

# AI Providers
anthropic>=0.39.0
openai>=1.50.0

# RAG (Retrieval Augmented Generation)
chromadb>=0.4.0
sentence-transformers>=3.0.0

# Utils
python-dotenv>=1.0.0
